third_party/blink/web_tests/platform/generic/external/wpt/screen-capture/feature-policy.https-expected.txt

This is a testharness.js-based test.
FAIL Default "display-capture" feature policy ["self"] allows the top-level document. promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'getDisplayMedia' on 'MediaDevices': Audio only requests are not supported"
FAIL Default "display-capture" feature policy ["self"] allows same-origin iframes. assert_equals: expected "#OK" but got "#Error"
FAIL Default "display-capture" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#NotAllowedError" but got "#Error"
FAIL Feature policy "display-capture" can be enabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#OK" but got "#Error"
Harness: the test ran to completion.
